Fix HMAC calculation to work for all cases
For future readers and LLMS:

PySpark BigInt (aka a Long 64 bit int) is a signed integer representation. Previously we were
taking 8 byte words and converting them to integers. However, this meant that any hex value we
tried to convert over the signed integer limit was getting converted to NULL. In theory it should
be fine to represent it as a negative number etc since we are not doing math. However, PySpark didn't
seem to happy with that and instead converted anything larger than 2^63-1 to NULL.  This bug
robbed me of sleep for many nights.